EES-2236 Move `DataGuidanceFileWriter` into Content.Services
As part of this, we have also removed the restriction for releases to
have `MetaGuidance` as this may not be possible for older releases where
there is no saved guidance.
