+++ src/GovUk.Education.ExploreEducationStatistics.Publisher/Services/DataGuidanceFileService.cs
@@ -3,11 +3,13 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common;
+using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Extensions;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Publisher.Services.Interfaces;
 using File = GovUk.Education.ExploreEducationStatistics.Content.Model.File;
 
@@ -32,16 +34,29 @@ namespace GovUk.Education.ExploreEducationStatistics.Publisher.Services
         public async Task<File> CreateDataGuidanceFile(Guid releaseId)
         {
             var path = Path.GetTempPath() + Guid.NewGuid() + ".txt";
-            await using var stream = await _dataGuidanceFileWriter.WriteFile(releaseId, path);
+
+            var release = await _contentDbContext.Releases.FindAsync(releaseId);
+
+            if (release is null)
+            {
+                throw new ArgumentException($"Could not find release with id: {releaseId}");
+            }
+
+            if (release.MetaGuidance.IsNullOrWhitespace())
+            {
+                throw new InvalidOperationException($"Release {releaseId} must have non-empty data guidance");
+            }
+
+            await using var stream = await _dataGuidanceFileWriter.WriteFile(release, path);
 
             var releaseFile = new ReleaseFile
             {
-                ReleaseId = releaseId,
+                Release = release,
                 Name = "Data guidance",
                 File = new File
                 {
                     Created = DateTime.UtcNow,
-                    RootPath = releaseId,
+                    RootPath = release.Id,
                     Filename = "data-guidance.txt",
                     Type = FileType.DataGuidance
                 }
