--- src/GovUk.Education.ExploreEducationStatistics.Publisher/Services/Interfaces/IDataGuidanceFileWriter.cs
@@ -1,11 +1,0 @@
-using System;
-using System.IO;
-using System.Threading.Tasks;
-
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Services.Interfaces
-{
-    public interface IDataGuidanceFileWriter
-    {
-        Task<FileStream> WriteFile(Guid releaseId, string path);
-    }
-}
